requirements.txt
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
brotlipy==0.7.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///home/builder/ci_310/cffi_1642753365720/work
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
commonmark==0.9.1
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588599/work
cycler==0.11.0
docker-pycreds==0.4.0
et-xmlfile==1.1.0
filelock==3.9.0
fonttools==4.34.4
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.9
GitPython==3.1.27
graphviz==0.20.1
huggingface-hub==0.11.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
joblib==1.1.0
kaleido==0.2.1
kiwisolver==1.4.4
kornia==0.6.9
lightning-bolts==0.6.0.post1
lightning-utilities==0.5.0
llvmlite==0.39.1
Markdown==3.4.1
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
mmcv-full==1.6.1
model-index==0.1.11
multidict==6.0.3
numba==0.56.4
numpy==1.23.1
opencv-python==4.6.0.66
openmim==0.2.0
openpyxl==3.0.10
ordered-set==4.1.0
packaging==21.3
pandas==1.4.3
pathtools==0.1.2
Pillow==9.2.0
plotly==5.11.0
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.12.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
pytorch-lightning==1.8.6
pytz==2022.1
PyYAML==6.0
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
rich==12.5.1
scikit-learn==1.1.1
scipy==1.8.1
sentry-sdk==1.9.0
setproctitle==1.3.0
shap==0.41.0
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
slicer==0.0.7
smmap==5.0.0
tabulate==0.8.10
tenacity==8.1.0
tensorboardX==2.5.1
threadpoolctl==3.1.0
timm==0.6.7
tokenizers==0.13.2
torch==1.12.0
torchaudio==0.12.0+cu116
torchcam==0.3.2
torchmetrics==0.11.0
torchsummary==1.5.1
torchvision==0.13.0+cu116
torchviz==0.0.2
tqdm==4.64.0
transformers==4.25.1
typing_extensions==4.3.0
urllib3==1.26.11
wandb==0.12.21
yapf==0.32.0
yarl==1.8.2
